(68.hrs48-56) Update order summary with the [ ] checkbox information
PSEUDOCODE FIRST

When order button is clicked, do these steps:

1. Get CHECKED STATE from CHECKBOX and store in a variable.
        To check if you're on the right path,  log the stored variable value to make sure that it reflects the actual checked state.
                Log.v("MainActivity", "The price is " + price);
                    By printing out the price here, we're verifying that after step one, the calculatePrice method does return the expected price.

2. PASS THIS INFORMATION into the order summary method.

3. Change the ORDER SUMMARY TEXT to include whether or not the user wants whipped cream.

-*-*-*-*3BOOP > Update Order Summary - Solution

PSEUDOCODE FIRST

When order button is clicked, do these steps:

1. Get CHECKED STATE from CHECKBOX and store in a variable.
        To check if you're on the right path,  log the stored variable value to make sure that it reflects the actual checked state.
                Log.v("MainActivity", "The price is " + price);
                    By printing out the price here, we're verifying that after step one, the calculatePrice method does return the expected price.

Step 1. indicates that we should add our code in the submitOrder()

Then we need to get the checked state from the checkbox.

In order to call any methods on a view, in the view hierarchy, we need to first find view by id.

The first thing I need to do is assign a view id to this check box view.

1.1. in activity_main.xml, add android:id="@+id/whipped_cream_checkbox"
1.2. in MainActivity.java, in submitOrder(), I start by creating a variable, whipedCreamCheckBox, with data type CheckBox. Then I can call the findViewById method within that activity:

CheckBox whippedCreamCheckBox = findViewById()

1.3. This will find a view based on the given ID; I'm going to pass in the ID which I just assigned in the other layout file.
(R.id.whipped_cream_checkbox)

This gives me an error immediately because the required type is CheckBox, because this is what the variable on the left-hand side is. However, the return value of this method is of type view.
So we need to cast the return value of this method.

I'm going to cast it to be a CheckBox object:
(CheckBox) findViewById(R.id.whipped_cream_checkbox);

The reason why I can cast it is because CheckBox is a subclass of View.

Ok, now I have the CheckBox Java object that represents the checkbox in the UI. In order to figure out whether the box is checked or not, I need to call a method on the object.

You may not know which method to call on the CheckBox object. And that's perfectly fine. So, GSearch:
checkbox android

The class overview is kind of interesting because it has some Java code here.

Let's try to use the method isChecked().

First we type the variable name of the CheckBox object. It's whippedCreamCheckbox.

Then I add a dot to call a method on it.

So now I go to the left side of this method call and then I'm going to store this result in a boolean variable:

boolean whippedCreamCheckBox.isChecked();

I'm going to call this boolean variable hasWhippedCream, then hit Enter:

boolean hasWhippedCream = WhippedCreamCheckBox.isChecked();

Now let's log this variable too, to make sure we're on the right path:

Log.v("MainActivity", "Has whipped cream: " + hasWhippedCream);

By concatenating this string here with the boolean variable, it will turn this boolean value into a string as well.

Now let's run the app and check it's actually printing out the right log message. Run, hit the checkbox, watch the log.