(69.hrs48-56) Pass the isChecked information into the orderSummary()
This is step 2 of pseudocode

I want to pass another argument into the createOrderSummary method, I'll just add a comma here, and then pass the variable, hasWhippedCream. Immediately I get an error that says, it can't be applied to the method createOrderSummary, because that just expects an integer parameter. So let's scroll down to where createOrderSummary is declared.

Now I need to modify the method signature of the createOrderSummary method. Instea of taking in one input parameter, it needs to take in two. So I'm going to add a comma and then specify the second parameter. The data type should be a boolean, and I'm going to call it addWhippedCream.

You can call the input parameter whatever name you want, but I just want to show you that this name does not have to match this name here (i.e. it does not have to match the hasWhippedCream param above.

Since we modified the method's signature, we have to update the Javadoc as well, to specify what this input parameter is about.

Ok, that completes the second task.

The third task is to change the order summary text to include whether or not the user wants whipped cream. Essentially that means we need to use this input parameter here, ("boolean addWhippedCream"), to modify our order summary price message.

In the screen shot, the whipped cram topping line appeared below the name. So let's add a new line here:

priceMessage += "Add whipped cream? " + addWhippedCream;

I also forgot to add the newline character in front:

pricemessage += "\nAdd whipped cream? " + addWhippedCream;

Save, run the app.