(78.hrs56-64) step3:
The code so far does pull out the name from the field.

Now that we have the user's name in a variable, and we verified that with logging, that it's the correct name, then we can update the order summary to show this name. In order for the order summary to show the correct name, we need to pass it as an input parameter to this method. That means that when we call the createOrderSummary method, we have to pass the name variable in as an input argument. Let's make that change first:

If I add "name" as an input argument, I get an error because the method call doesn't match the method signature. Scrolling down to the method definition, I can add String name as an input parameter. I'll also update the Javadoc as well.

Now, inside the createOrderSummary method, I can use this variable. I'm going to erase the hard coded name and then concatenate this name variable.