(80.hrs56-64) Explanation of this line: String name = nameField.getText().toString();
This bit of code, getText(), returns the return data type "Editable".
So this bit of code, nameField.getText(), returns "Editable" data type.
This returns an editable object, and we can call another method on this object by adding the dot notation and then the method name.

So in this case we're calling the toString() method on the editable object, and the toString method returns a string, so that can be stroed in the name variable. This is a more advanced concept, and it's called chaining method calls. The reason we are doing this here is because the getText() method returns an editable object. An editable object cannot be stored inside a string variable, as seen by this Android Studio error saying incompatible types. So if we add the toString()
method, then the return value of this whole thing is a string, and that can be stored in the name variable.