(80.hrs56-64) ADJUST PRICE BASED ON TOPPINGS
1. Calculate base price of 1 cup
$5 for coffee + $1 for whipped cream = $6 per cup

2. Find total order price
$6 per cup * 3 cups

PLAN OUT YOUR STEPS FIRST

Then implement this in your app.
What method did you modify?

In order to calculate the correct price for the order summary, we should pass the information about which toppings the user has selected into the calculatePrice metod.

By consolidating everything within this method, if we have a problem with how the price is calculated, we can check the code in a single place.

It's a lot more manageable.

Then within this method,

we're going to figure out the base price of one cup of coffee.

This is just like I showed you earlier.

If the coffee should have whipped cream, then add a dollar.

If the coffee should have chocolate, then add $2.

Then take the base price of one cup of coffee and multiply it by  the quantity ordered to determine the total order price.

CODE:

(81) First let's pass in the information about the whipped cream and chocolate check boxes into the calculatePrice method. I'm going to pass in the booleans hasWhippedCream and hasChocolate.