(86.hrs64-72) Solving the negative number of cups of coffee.
There also should be a max number of coffees.

We can do this by INPUT CHECKING.

For this quiz, make this change in your app:
Only allow the user to order 1-100 cups of coffee.

Assume that the app always starts at two cups of coffee when you first launch it.

To find the right place to make the code change, think about when the problem first occurs.

What button does the user press, and what code is triggered by that button?

Since the app has different behavior depending on whether the number of cups of coffee is valid or not, think about using a control flow statement like an if else statement.

That way, the logic in the app can be different based on a certain condition. If the user is trying to order a number of cups of coffee outside of these bounds, then figure out how to stop the app from continuing.

Once you make these changes, answer these questions.

1. Which method did you modigy so the quantity doesn't go below 1 cup of coffee?

2. Which method did you modify so the quantity doesn't go above 100 cups of coffee?

As a hint, it may be useful for you to know that you can put a return statement anywhere inside a method and that will cause you to exit the method early. Then the rest of the code after that will be skipped.

Challenge, extend the app to also add a popup toast message that explains to the user why the quantity insn't updating when it hits these bounds.

By the way, if you want to test that the upper bound works, instead of tapping the plus sign 100 times, you can just make a temporary change in the code. You can change the initial global quantity variable in the Java file and the initial quantity value in the XML
to start at a higher number like 99 instead of just 2.

CHECK INPUT BOUNDS

Plan:
When the user tries to order too FEW coffees, what METHOD
will you need to modify?
    - decrement()

What CONDITION will you use?
    - if(quantity <1) {...

When the user tires to order too MANY coffees, what method will you need to modify?
    - increment()

What condition will you use?
    - if (quantity > 100)

We're going to start by looking at the lower bound first.

 Remember that the minus button triggers the decrement method in the Java file.

 So we need to modify the decrement method and prevent the app from updating the quantity on the screen.

 I open up the code in the Java file for the decrement method.

 Once we're inside this method, we don't want this line to be executed if the quantity is already 1:

 quantity = quantity - 1;

Otherwise, this line of code is going to cause a quantity to change to 0.

I'm gonna insert an if statement at the top of the method:

public void decrement (View view) {
    if (quantity == 1) {
        // if the quantity is not 1, then we'll just skip this
        // whole block of code and continue updating the quantity
        // after the code block,
        // and if it is, then
        // *STOP COUNTING LOWER*
    }
    quantity = quantity - 1;
    displayQuantity(quantity);
}

How I'll stop counting inside the if statment?

I'll write the return statement!

The return statement will stop updating the quantity = quantity - 1; and jump out of the decrement method.