(87.hrs64-72) Now let's look at the case with the upper bound.
The only way to hit 100 is by pressing the plus button when there is already 100 cups of coffee.

The plus button triggers the increment method.

So we need to modify this method to prevent the app from updating the quantity to a 101 cups of coffee.

Within the increment method we add some additional logic like we did for the decrement method.

If the quantity variable is currently at a 100 then we should exit the method early. Otherwise we should continue by incrementing the quantity by one more cup of coffee and then displaying it on our screen. Let's run this on our device to see what happens.

In order to test that the upper bound quantity works, I'm gonna change the quantity variable in our Java file to be 99 cups of coffee.

I'm gonna make a similar change in the layout as well.