(88.hrs64-72)  // Show an error message as a toast Toast.makeText(this, "You cannot have more than 100 coffees", Toast.LENGTH_SHORT).show(); // Exit this method early because there's nothing left to do return;
To create a toast message, you call Toast.makeText and you pass in three different input arguments.

You pass in a context, which is "this" activity, you pass it a message, "you cannot...", and you pass in how long it should show up on screen.

All this code creates a single toast object, and then we call a method show on that object.

We add a similar toast to the decrement method.