(91.hrs64-72) Send off an order summary in an email
It opens up an email app on the device with a prepopulated message containing the coffee order and a custom subject line.

In order to accomplish this we need to understand a new topic in Android called intents.

An intent is a message that requests some type of action to be performed by another app component, such as an activity in another app.

"An intent is a call to action!"

This is a powerful functionality, because our app can use the functionality provided by another app, like sending an email, opening the camera, reviewing a map.

It lets us be efficient and smart as developers, because we're not reinventing what other apps already do well.

Think of SENDING INTENTS like tossing a ball to someone.

E.g. in the Yelp app, clicking on the location of a business sends an intent to open up a maps app on the device to provide turn by turn navigation instructions to the user.

The coolest part about sending intents is that you don't have to know precisely which app is capable of handling it. There could be multiple apps on the device that could handle the intent, but ultimately only one needs to handle it. Using our analogy, when you throw the ball, you don't have to know who you're throwing it at, as long as someone catches it.

If multiple apps, the popup window, known as disambiguation dialogue. Now let's move on to talking about what's contained inside this ball that represents an intent.

You can think of an intent as a message specifying some operation to be performed. It contains two primary things, the general action and some data to act upon.

You can also specify additional information such as category, component, and extras, which is literally extra information. All of these things are used by Android to determine which app component can actually handle the intent. For example a maps intent would contain the action ACTION_VIEW, and the data would be the latitude and longitude coordinates of the location that we want to see.

The rest of the fields can be blank. Then, any app that's capable of displaying latitude and longitude coordinates can be used to handle the request, which is probably any maps application. Another example is to use an intent to dial a phone number. In this case, you would compose an intent where the action is ACTION_DIAL and the corresponding data is the telephone number to call. Notice that the format is in the form tel, followed by the number. This formatted string is called a URI, short for uniform resource identifier.

This is a standardized way to structure data so that it can be properly processed on the receiving end.

Think about it like a mailing address, it follows a structure so that it can be delivered properly. In case of sending an intent, a URI is needed so that the data can be properly passed to the receiving app component. Depending on the data encoded in the intent the scheme may change, such as this case or the geo case that we saw earlier.

-*-*-*-*-*3B >> Practice With Intents

EMAIL OUT THE ORDER SUMMARY

Task: Use an intent to send an order summary to an email app.

But first take some time to experiment with writing code for different intents listed in the Common Intents guide.

http://developer.android.com/guide/components/intents-common.html

It shows what's possible on Android.

CTRL / to comment out the whole selected text on submitOrder() method signature

When the order button is clicked, I want to send a maps intent.

Paste the map intent:

Intent intent = new Intent(Intent.ACTION_VIEW);
    intent.setData(geoLocation); // 1
    if (intent.resolveActivity(getPackageManager()) != null) {
        startActivity(intent);
    }

In this line of code (// 1), we're creating a new Intent object using the "new" keyword in Java. We're calling the Intent constructor, and passing it a ACTION_VIEW. A constant is a variable whole contents, once set, never changes. ACTION_VIEW or ACTION_SEND are defined as constants within the Intent class, hence: Intent.ACTION_VIEW.

So to recap, the // 1 line of code creates a new object instance called intent. And then on the Intent object we can call different methods like setData. We want to set the data URI to be a certain set of latitude and longitude coordinates. Currently the variable geoLocation is unrecognized in our app, so I'm going to replace it.

The replacement code creates a Uri from this string here. And then this Uri is set as data field within this intent object.

Now we have our maps intent created.

The next two lines down here, we have some code that looks a little bit confusing. There's a lot going on, but essentially we're just preventing the app from crashing by checking to see if the Intent can be handled.

In terms of our ball analogy, it means that if no one is available to catch the ball, then the ball drops and there's an explosion because the system doesn't know what to do next, and that's why we need this line of code here.

We're checking to see if there's at least one component on the device that can handle this Intent. If so, then this condition for the if statement is true, and then we execute the line of code here within these braces.

We go ahead and start the activity with this Intent. If there's no app component on the device that can handle this Intent, then the test condition is false. And we skip this block of code and we do not send the Intent.

After you finish experimenting with Intents, go ahead and extend your app to have this functionality:

EMAIL OUT THE ORDER SUMMARY
Experiment with sending intents from the Common Intents guide.