(92.hrs64-72) [1] Use an intent to send the order summary to an email app.
[2] Populate email message with this text.

[3] Remove code that displays the order summary in the Just_Java app.

GSEARCH: email intent android

Copy and paste the code snippet:

    Intent intent = new Intent(Intent.ACTION_SENDTO);
    intent.setData(Uri.parse("mailto:")); // only email apps should handle this
    intent.putExtra(Intent.EXTRA_EMAIL, addresses);
    intent.putExtra(Intent.EXTRA_SUBJECT, subject);
    if (intent.resolveActivity(getPackageManager()) != null) {
        startActivity(intent);
    }

But where to paste the code snippet? Well, we want to send the order summary off to the email app as soon as we have the complete order summary. And the place that we have the complete order summary is after this line of code here which creates the order summary and then stores it in a string variable called priceMessage.

So I'm going to add a new line here, and then insert our code.

Notice how we're creating a new intent object using the constructor with the new keyword in java. We're passing in one argument as input into the constructor to initialize this intent object. The argument is this constant called Intent.ACTION_SENDTO. A constant is a variable who's contents should never change once it's set. In this case when we're creating our intent we're using official intent action. ACTION_SENDTO
which is defined as a constant in the intent class. Then we call some methods on this newly created intent object. We set the data to be a URI of this format, "mailto:", that way only email apps will handle this intent. Then we add some extra data into the intent. We don't need to specify the two field of this e-mail so I'm just going to delete that line. We do, however, want to specify the subject of the e-mail message. So I'm going to replace this with actual subject line. "Just_Java order for " + name.
Then these next couple of lines of code check if there's an activity on the device to handle this intent. If there is then it's safe to send the intent and recall start activity using this intent.
If there's no activity on the device to handle this intent, then we skip this block of code and we skip sending the intent.

Now let's run the app on our device. It sends an intent to an email app on the device.