(93.hrs64-72) Now, the original quiz question asked us to populate the email subject line, as well as the message body with the order summary. So let's go do that now. Looking back at the documentation for the email intent, I'm going to scroll up to look at the intent extras. We need to add an intent extra for each additional piece of information we want to provide about the email. For example, we could have an intent extra about the recipients of the email, and we could have an Intent.EXTRA about the subject line of the email. But this is what we really want:
We want ot add an Intent.EXTRA to specify the body of the email. Again, Intent.EXTRA_TEXT is a constant defined in the intent java class. If I click on it you'll see more information. It's defined as a String in the intent class. And the constant value is this string right here:
"Constant Value: 'android.intent.extra.TITLE' "

In the Just Java app when we create the intent and we're going to populate this extra, containing the body of the email message.

When the email app receives the intent, they'll know to check this intent extra to find the body of email and then they can populate that into the user interface.

Let's make that code change now. I'm going to select this line of code,

"
intent.putExtra(Intent.EXTRA_SUBJECT, "Just_Java order for " + name);
"

...that puts one extra into the intent for the subject of the email. On the next line, I'm going to paste the code. Now we're putting a second extra into the intent. The extra is going to correspond to Intent.EXTRA_TEXT, so that we know it's for the body of the email. The second argument to this method is going to be the actual body of the email. The body of the email should just containt the order summary and that's stored in price message so I can just copy this variable over to here:

"
intent.putExtra(Intent.EXTRA_SUBJECT, priceMessage);
"

To recap, for the intent object we're calling a method putExtra to spcify the email subject. IN THE SAME INTENT OBJECT, on the next line, we're calling the put extra method again, to specify the body of the email.

And the body of the email is the order summary. Now let's run the app on our device. Once I complete the order, I can click the button tu submit. We send an intent to the email app with the email subject populated and the order summary shows up now.