(94.hrs64-72) For the third step of this task we're going to remove the code that displays the order summary within the Just Java app because now it's gonna show up in the email app. I'm going to start by modifying the activity main layout to remove the order summary text views. At the bottom of the layout, I'm going to delete the header, as well as the Text View that would show the actual order summary.
I'm gonna save that and switch to the Java file. In the main acitivity Java file, I no longer need this display message method, which would update the order summary text view because this view with this ID
doesn't exist anymore.

Since I deleted the display message method, Android studio now gives me an error saying that this method doesn't exist, so I'm going to delete this:
        displayMessage(priceMessage);

Looking back at our code, this makes sense, because when we click on the order button, it's going to call the submit order method, we're going to do all this code, which also sends an intent to the email app. There's no need to display the order summary within our app any more. Let's verify that the app still works.